% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ifelse_null}
\alias{ifelse_null}
\title{Returns NA if input is null, else returns input}
\usage{
ifelse_null(x)
}
\arguments{
\item{x}{A number.}
}
\value{
A number or NA.
}
\description{
Returns NA if input is null, else returns input
}
\examples{
ifelse_null(NA)
ifelse_null(1)
}
