% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{item_base_response}
\alias{item_base_response}
\title{Returns tibble of items in base response format}
\usage{
item_base_response(x)
}
\arguments{
\item{x}{A List.}
}
\value{
A tibble with 15 columns in base response format.
}
\description{
response formats are described at the url
\url{https://developer.walmartlabs.com/docs/read/Item_Field_Description}.
}
