% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{save_walmart_credentials}
\alias{save_walmart_credentials}
\title{Save API credentials for later use}
\usage{
save_walmart_credentials(app_key)
}
\arguments{
\item{app_key}{application key}
}
\description{
This functions caches the credentials to avoid need for entering it when calling other functions
}
\examples{
# since not checking is preformed not to waste API calls
# it falls on the user to save correct information
save_walmart_credentials("APP_KEY")
}
