% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{smart_subset}
\alias{smart_subset}
\title{Subsets up to n elements in a vector}
\usage{
smart_subset(x, n)
}
\arguments{
\item{x}{A vector.}

\item{n}{A number.}
}
\value{
vector of length equal to length of x or n, whichever is smallest.
}
\description{
Subsets up to n elements in a vector
}
\examples{
smart_subset(1:10, 5)
smart_subset(1:10, 50)
}
