% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending.R
\name{trending}
\alias{trending}
\title{Trending products at Walmart.com}
\usage{
trending(key = auth_cache$KEY, lsPublisherId = NULL, list_output = FALSE)
}
\arguments{
\item{key}{Your API access key.}

\item{lsPublisherId}{Your LinkShare Publisher Id.}

\item{list_output}{Indicator for list output.}
}
\value{
A tibble with 15 columns in base response format.
}
\description{
\code{\link{trending}} gives information on what is bestselling on
Walmart.com right now.
}
\details{
An API key will be required to run this function and can be acquired by
creating an account on the following website
\url{https://developer.walmartlabs.com/member}.

For more information refer to the original documentation
\url{https://developer.walmartlabs.com/docs/read/Trending_API}.

Response formats are described at the url
\url{https://developer.walmartlabs.com/docs/read/Item_Field_Description}.
}
\examples{
\dontrun{
key <- "************************"

trending(key = key)

trending(key = key, list_output = TRUE)
}
}
