% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vod.R
\name{VOD}
\alias{VOD}
\title{Value of the day}
\usage{
VOD(key = NULL, list_output = FALSE)
}
\arguments{
\item{key}{Your API access key.}

\item{list_output}{Indicator for list output.}
}
\value{
A tibble with 15 columns in base response format.
}
\description{
\code{\link{VOD}} provides the Value of the Day item at walmart.
}
\details{
An API key will be required to run this function and can be aqquired by
creating an account on the following website
\url{https://developer.walmartlabs.com/member}.

Response formats are described at the url
\url{https://developer.walmartlabs.com/docs/read/Item_Field_Description}.
}
\examples{
\dontrun{
key <- "************************"

VOD(key = key)
}
}
